refactor: apply best practices and fixes
- replace panic with error handling in config
- change ClickCount to int64 for DB consistency
- remove unused db.go service
- fix error messages formatting in repository
- add proper error handling in main.go

+++ internal/model/url.go
@@ -8,5 +8,5 @@ type URL struct {
 	OriginalURL string
 	CreatedAt   time.Time
 	ExpiresAt   *time.Time
-	ClickCount  int
+	ClickCount  int64
 }
